make game-engine app more thin
